# pkm_frontend/libs/api/model/aIRequest.ts
/**
 * PKM API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AIRequest { 
    /**
     * The user\'s message to the AI, asking to create a note
     */
    message: string;
    /**
     * The folder where the note should be placed (folder ID)
     */
    folder: string;
    /**
     * Flag to indicate if the AI should generate a note from the message
     */
    createNoteIfPossible?: boolean;
}